pyproject.toml
[project]
name = "attestor"
version = "4.0.0"
description = "Production-grade memory infrastructure for multi-agent systems. Namespace isolation, RBAC, provenance, ranked retrieval."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10,<4.0"
authors = [{ name = "aarjay" }]
keywords = ["ai", "agent", "memory", "multi-agent", "llm", "rbac", "namespace", "provenance", "postgres", "pgvector", "neo4j", "graph", "arangodb", "mcp", "agentic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mcp>=1.0.0",
    "tomlkit>=0.12.0",
    "starlette>=0.27.0",
    "uvicorn>=0.23.0",
    # Postgres + Neo4j are the v4 default backends. The codebase imports
    # them eagerly (attestor/conversation/episodes.py, store/postgres_backend.py,
    # store/neo4j_backend.py), so they must be present at install time for
    # `import attestor.conversation` to even resolve. Cloud-only deployments
    # that swap to ArangoDB / DynamoDB / Cosmos / AlloyDB will install
    # these as harmless extras until backend imports are made lazy.
    "psycopg2-binary>=2.9.0",
    "neo4j>=5.0.0",
    # `requests` is used by the OllamaEmbeddingProvider — the documented
    # default embedder. Without it, `attestor doctor` fails on a fresh
    # install with "No embedding provider available" even when Ollama is
    # running locally. Lazy-imported in store/embeddings.py and
    # store/aws_backend.py.
    "requests>=2.28.0",
]

[project.urls]
Homepage = "https://attestor.dev"
Repository = "https://github.com/bolnet/attestor"
Issues = "https://github.com/bolnet/attestor/issues"

[project.optional-dependencies]
extraction = ["openai>=1.0.0"]
arangodb = ["python-arango>=8.0.0"]
postgres = ["psycopg2-binary>=2.9.0"]
neo4j = ["neo4j>=5.0.0"]
aws = ["boto3>=1.28.0"]
azure = ["azure-cosmos>=4.5.0", "azure-identity>=1.14.0", "azure-search-documents>=11.4.0", "networkx>=3.0"]
gcp = ["google-cloud-firestore>=2.11.0", "google-cloud-aiplatform>=1.38.0"]
cloud-embeddings = ["openai>=1.0.0"]
voyage = ["voyageai>=0.3.0; python_version < '3.14'"]
lambda = ["mangum>=0.17.0", "starlette>=0.27.0"]
docker = ["docker>=6.0"]
eval = ["braintrust>=0.0.150", "anthropic>=0.30.0", "PyYAML>=6.0"]
all = [
    "openai>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "python-arango>=8.0.0",
    "neo4j>=5.0.0",
    "boto3>=1.28.0",
    "azure-cosmos>=4.5.0",
    "azure-identity>=1.14.0",
    "azure-search-documents>=11.4.0",
    "networkx>=3.0",
    "google-cloud-firestore>=2.11.0",
    "google-cloud-aiplatform>=1.38.0",
    "docker>=6.0",
]

[project.scripts]
attestor = "attestor.cli:main"

[tool.poetry]
packages = [{ include = "attestor" }]
# Externalized prompt templates ship as .md files next to the loader.
# Poetry includes data files inside packages by default once an
# __init__.py exists, but listing them here makes the contract explicit
# and survives any future packaging-format changes.
include = [
    { path = "attestor/extraction/prompts/*.md", format = ["sdist", "wheel"] },
]

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0"
pytest-cov = ">=4.0"
pytest-asyncio = ">=0.23"
requests = ">=2.28.0"
twine = "^6.2.0"
ruff = ">=0.6"
mypy = ">=1.10"
hypothesis = ">=6.100"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF", "ANN", "ASYNC", "S", "PT", "RET", "TCH", "FA"]
ignore = [
    "ANN101",  # missing type for self
    "ANN102",  # missing type for cls
    "ANN401",  # Any in signatures (legitimate at boundaries)
    "S101",    # assert (allowed in tests)
    "S603",    # subprocess (we control inputs)
    "S607",    # subprocess partial-path
    "B008",    # function call in default arg (FastAPI/Starlette pattern)
    "RET504",  # unnecessary assign-then-return (often clearer)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "S", "PLR2004", "RUF"]
"attestor/longmemeval*.py" = ["BLE001"]   # eval harness intentionally swallows
"attestor/locomo.py" = ["BLE001"]
"attestor/cli.py" = ["T201"]              # CLI prints
"attestor/doctor_v4.py" = ["T201"]        # doctor prints diagnostics
"attestor/init_wizard.py" = ["T201"]      # interactive wizard prints
"attestor/_branding.py" = ["T201"]
"experiments/*" = ["ANN", "T201", "BLE001"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unreachable = true
warn_redundant_casts = true
files = [
    "attestor/__init__.py",
    "attestor/context.py",
    "attestor/models.py",
    "attestor/client.py",
]
# expand `files` module-by-module as the codebase tightens

[[tool.mypy.overrides]]
module = ["psycopg2.*", "neo4j.*", "pinecone.*", "voyageai.*", "boto3.*", "azure.*", "google.*", "mcp.*", "starlette.*", "uvicorn.*", "tomlkit.*", "networkx.*"]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "docker: tests requiring Docker (deselect with '-m not docker')",
    "unit: fast, dependency-free unit tests",
    "integration: tests that touch real services or the filesystem beyond fixtures",
    "live: tests that hit the live attestor_v4_test Postgres DB and clobber its schema; deselect during bench runs with '-m \"not live\"'",
]
